+++ js/templates/init-replay.js
@@ -1,6 +1,6 @@
 import {initialState, currentPlayer} from '../data/game.js';
 import showScreen from './show-screen.js';
-import getScreenWelcome from './screens/welcome.js';
+import getScreenWelcome from './screens/welcome/welcome.js';
 
 const initReplay = () => {
   const replay = document.querySelector(`.js-main-replay`);
