+++ js/templates/screens/welcome/welcome.js
@@ -1,0 +1,14 @@
+import controlGame from '../../control-game.js';
+import WelcomeView from './welcome-view.js';
+
+const getScreenWelcome = (state) => {
+  const screenWelcome = new WelcomeView();
+
+  screenWelcome.onPlayButtonClick = () => {
+    controlGame(state);
+  };
+
+  return screenWelcome.element;
+};
+
+export default getScreenWelcome;
